pyproject.toml
[tool.poetry]
name = "image-search"
version = "0.1.0"
description = "A fastapi web app that uses multi modal similarity to search for images of clothes webscrapped from blogshops."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.*"
requests = "2.31.0"
beautifulsoup4 = "4.12.3"
selenium = "4.17.2"
pandas = "2.2.0"
numpy = "1.26.4"
chromadb = "0.4.24"
open-clip-torch = "2.24.0"
torch = "2.2.1"
torchvision = "0.17.1"
tabulate = "0.9.0"
python-dotenv = "1.0.1"
psycopg2-binary = "2.9.9"
sqlalchemy = "1.4.52"
fastapi = "0.110.1"
pre-commit = "3.7.0"
flake8-pyproject = "1.2.3"
uvicorn = "0.27.0"
protobuf = "3.20.*"
webdriver-manager="4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.2"
pytest-cov = "5.0.0"
isort = "5.13.2"
black = "24.2.0"
flake8 = "7.0.0"
bandit = "1.7.8"
mypy = "1.9.0"
interrogate = "1.7.0"
httpx = "0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 150
include = '\.pyi?$'
exclude = '''
/(
      .eggs         # exclude a few common directories in the
    | .git          # root of the project
    | .hg
    | .mypy_cache
    | .tox
    | venv
    | _build
    | buck-out
    | build
    | dist
  )/
'''

[tool.isort]
profile = "black"
line_length = 79
multi_line_output = 3
include_trailing_comma = true
virtual_env = "venv"

[tool.flake8]
exclude = ".venv"
ignore = ["E501", "W503", "E226"]
# E501: Line too long
# W503: Line break occurred before binary operator
# E226: Missing white space around arithmetic operator

[tool.bandit]
exclude_dirs = [".venv"]
tests = [] #"B201", "B301"
skips = ["B101"] #, "B601"

[tool.pytype]
inputs = ['package_name']

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-module = true
ignore-property-decorators = true
fail-under = 80
exclude = ["tests/", ".venv/"]
verbose = 0
omit-covered-files = false
quiet = false
whitelist-regex = []
ignore-regex = []

[tool.mypy]
python_version = "^3.11"
warn_return_any = true
warn_unused_configs = true
